+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## [1.2.3](https://github.com/adobe/da-collab/compare/v1.2.2...v1.2.3) (2026-05-05)
+
+
+### Bug Fixes
+
+* avoid storage.deleteAll() in storeState to prevent Cloudflare DO reset bug ([#133](https://github.com/adobe/da-collab/issues/133)) ([4ebbddc](https://github.com/adobe/da-collab/commit/4ebbddc51048109ceb96e212d1d9397cdd9787ee))
+
+## [1.2.2](https://github.com/adobe/da-collab/compare/v1.2.1...v1.2.2) (2026-05-05)
+
+
+### Bug Fixes
+
+* skip stale da-admin reload when client pushes Y.js state after reconnect ([#136](https://github.com/adobe/da-collab/issues/136)) ([a8ddc23](https://github.com/adobe/da-collab/commit/a8ddc23256f91e8e020b6d34ca5fdd6e39c37f45))
+
 ## [1.2.1](https://github.com/adobe/da-collab/compare/v1.2.0...v1.2.1) (2026-04-30)